# src/modules/product/group/product-group.controller.ts
import {
  BadRequestException,
  Controller,
  Get,
  Query,
  Req,
  UseGuards,
} from '@nestjs/common';
import {
  JwtAuthGuard,
  type ReqWithAuthContext,
} from 'src/modules/auth/guards/jwt-auth.guard';
import {
  ApiBearerAuth,
  ApiOperation,
  ApiQuery,
  ApiResponse,
} from '@nestjs/swagger';
import { ProductGroupService } from './product-group.service';
import { PermissionsGuard } from 'src/infra/rbac/permissions.guard';
import { RequirePermissions } from 'src/infra/rbac/permissions.decorator';

import { ProductGroupResponseDto } from '../dto/product-response.dto';

@Controller('products/groups')
export class ProductGroupController {
  constructor(private readonly groupService: ProductGroupService) {}

  @Get()
  @UseGuards(JwtAuthGuard, PermissionsGuard)
  @RequirePermissions({ allOf: ['tenant.groups.view'] })
  @ApiBearerAuth()
  @ApiOperation({
    summary: 'Listar grupos de produtos',
    description:
      'Retorna uma lista paginada de grupos de produtos associadas às credenciais do usuário autenticado.',
    tags: ['Product / Group'],
  })
  @ApiResponse({
    status: 200,
    description: 'Lista de grupos de produtos retornada com sucesso.',
    type: [ProductGroupResponseDto],
  })
  @ApiResponse({
    status: 400,
    description: 'Erro de requisição, como pageSize excedendo o limite.',
  })
  @ApiResponse({
    status: 401,
    description: 'Token de autenticação inválido ou ausente.',
  })
  @ApiQuery({
    name: 'page',
    required: false,
    type: Number,
    description: 'Número de página para paginação',
  })
  @ApiQuery({
    name: 'pageSize',
    required: false,
    type: Number,
    description: 'Número de itens por página',
  })
  getBrands(
    @Req() req: ReqWithAuthContext,
    @Query('page') page?: number,
    @Query('pageSize') pageSize?: number,
  ) {
    const credentialsId = req.authContext?.credentialsId;

    if (!credentialsId) {
      throw new Error('Credentials ID not found in token');
    }

    return this.groupService.get(credentialsId, page, pageSize);
  }
}