# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/products/brands": {
      "get": {
        "description": "Retorna uma lista paginada de marcas de produtos associadas às credenciais do usuário autenticado.",
        "operationId": "ProductBrandController_getBrands",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de página para paginação",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Número de itens por página",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de marcas de produtos retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductBrandResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de requisição, como pageSize excedendo o limite."
          },
          "401": {
            "description": "Token de autenticação inválido ou ausente."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar marcas de produtos",
        "tags": [
          "Product / Brand"
        ]
      },
      "post": {
        "description": "Cria uma nova marca de produto associada ao tenant.",
        "operationId": "ProductBrandController_createBrand",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBrandDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Marca criada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductBrandResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação ou de requisição."
          },
          "401": {
            "description": "Token de autenticação inválido ou ausente."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar marca de produto",
        "tags": [
          "Product / Brand"
        ]
      }
    },
    "/api/v1/products/brands/{id}": {
      "patch": {
        "description": "Atualiza os dados de uma marca de produto existente no tenant.",
        "operationId": "ProductBrandController_updateBrand",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Código da marca (MAR_CODIGO) a ser atualizada",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBrandDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Marca atualizada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductBrandResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação ou de requisição."
          },
          "401": {
            "description": "Token de autenticação inválido ou ausente."
          },
          "404": {
            "description": "Marca não encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar marca de produto",
        "tags": [
          "Product / Brand"
        ]
      }
    },
    "/api/v1/products/groups": {
      "get": {
        "description": "Retorna uma lista paginada de grupos de produtos associadas às credenciais do usuário autenticado.",
        "operationId": "ProductGroupController_getBrands",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Número de página para paginação",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "Número de itens por página",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de grupos de produtos retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductGroupResponseDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de requisição, como pageSize excedendo o limite."
          },
          "401": {
            "description": "Token de autenticação inválido ou ausente."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar grupos de produtos",
        "tags": [
          "Product / Group"
        ]
      }
    },
...
  },
  "info": {
    "title": "InfoBrasil API",
    "description": "API de integração do sistema Retaguarda (by InfoBrasil) com outros sistemas, como ERPs, e-commerce, etc. <br><br>\n      Se deseja solicitar um acesso, preencha o formulário em: <a href='https://info-api.infobrasilsistemas.com.br/integration/form'>Solicitar Acesso</a>\n      <br><br>\n      **O sistema conta com duas formas de autenticação:**<br> \n      - Basic Auth para login e obtenção de token JWT<br>\n      - JWT para acesso aos endpoints protegidos.\n      <br><br>\n      Para verificar a disponibilidade da API acesse nossa <a href='/status'>Página de Status</a>\n      <br><br>\n      A API é organizada em módulos, cada um responsável por uma área específica do sistema, como produtos, clientes, vendas, etc.",
    "version": "1.15.1",
    "contact": {
      "name": "InfoBrasil Sistemas",
      "url": "https://www.infobrasilsistemas.com.br",
      "email": "suporte@infobrasilsistemas.com.br"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3336",
      "description": "Servidor local para desenvolvimento"
    },
    {
      "url": "https://info-api.infobrasilsistemas.com.br",
      "description": "Servidor de produção"
    }
  ],
  "components": {
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProductBrandResponseDto": {
        "type": "object",
        "properties": {
          "MAR_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código da marca"
          },
          "MAR_DESCRICAO": {
            "type": "string",
            "example": "Marca Exemplo",
            "description": "Descrição da marca"
          }
        },
        "required": [
          "MAR_CODIGO",
          "MAR_DESCRICAO"
        ]
      },
      "CreateBrandDto": {
        "type": "object",
        "properties": {
          "MAR_DESCRICAO": {
            "type": "string",
            "description": "Descrição da marca",
            "example": "Marca Exemplo"
          }
        },
        "required": [
          "MAR_DESCRICAO"
        ]
      },
      "UpdateBrandDto": {
        "type": "object",
        "properties": {
          "MAR_DESCRICAO": {
            "type": "string",
            "description": "Descrição da marca",
            "example": "Marca Atualizada"
          }
        }
      },
      "ProductGroupResponseDto": {
        "type": "object",
        "properties": {
          "GRU_CODIGO": {
            "type": "number",
            "example": 1,
            "description": "Código do grupo"
          },
          "GRU_DESCRICAO": {
            "type": "string",
            "example": "Grupo Exemplo",
            "description": "Descrição do grupo"
          }
        },
        "required": [
          "GRU_CODIGO",
          "GRU_DESCRICAO"
        ]
      },
...
    }
  }
}